change ResourceManager to fileManager
