+++ superagi/helper/google_calendar_creds.py
@@ -10,7 +10,7 @@ from googleapiclient.discovery import build
 from sqlalchemy.orm import sessionmaker
 from superagi.models.db import connect_db
 from superagi.models.tool_config import ToolConfig
-from superagi.resource_manager.manager import ResourceManager
+from superagi.resource_manager.manager import FileManager
 
 class GoogleCalendarCreds:
 
@@ -28,7 +28,7 @@ class GoogleCalendarCreds:
             engine = connect_db()
             Session = sessionmaker(bind=engine)
             session = Session()
-            resource_manager: ResourceManager = None
+            resource_manager: FileManager = None
             with open(file_path,'rb') as file:
                 creds = pickle.load(file)
             if isinstance(creds, str):
