+++ superagi/jobs/agent_executor.py
@@ -22,7 +22,8 @@ from superagi.models.project import Project
 from superagi.models.tool import Tool
 from superagi.models.tool_config import ToolConfig
 from superagi.tools.base_tool import BaseToolkitConfiguration
-from superagi.resource_manager.manager import ResourceManager
+from superagi.resource_manager.manager import FileManager
+from superagi.resource_manager.resource_manager import ResourceManager
 from superagi.tools.thinking.tools import ThinkingTool
 from superagi.tools.tool_response_query_manager import ToolResponseQueryManager
 from superagi.vector_store.embedding.openai import OpenAiEmbedding
@@ -262,7 +263,7 @@ class AgentExecutor:
             if hasattr(tool, 'agent_id'):
                 tool.agent_id = agent_id
             if hasattr(tool, 'resource_manager'):
-                tool.resource_manager = ResourceManager(session=session, agent_id=agent_id)
+                tool.resource_manager = FileManager(session=session, agent_id=agent_id)
             if hasattr(tool, 'tool_response_manager'):
                 tool.tool_response_manager = ToolResponseQueryManager(session=session, agent_execution_id=parsed_config[
                     "agent_execution_id"])
