+++ superagi/tools/file/write_file.py
@@ -3,7 +3,7 @@ from typing import Type, Optional
 from pydantic import BaseModel, Field
 
 # from superagi.helper.s3_helper import upload_to_s3
-from superagi.resource_manager.manager import ResourceManager
+from superagi.resource_manager.manager import FileManager
 from superagi.tools.base_tool import BaseTool
 
 
@@ -31,7 +31,7 @@ class WriteFileTool(BaseTool):
     args_schema: Type[BaseModel] = WriteFileInput
     description: str = "Writes text to a file"
     agent_id: int = None
-    resource_manager: Optional[ResourceManager] = None
+    resource_manager: Optional[FileManager] = None
 
     class Config:
         arbitrary_types_allowed = True
